UPSTREAM:DP:wifi:ath12k:fix paddr truncation in ath12k_dp_link_desc_init
The local variable 'paddr' in ath12k_dp_link_desc_init() was
declared as u32, which causes silent truncation of physical
addresses on platforms with 64-bit DDR where memory can be
mapped above the 4GB boundary.

Fix this by changing the type of 'paddr' from u32 to u64 to
correctly hold the full 64-bit physical address returned during
link descriptor initialization.

Fixes: address truncation on systems with >4GB DDR

Patch-dependency: None

(cherry picked from commit 9f150046311242b75fbad8683b9350e7066fcbc1)

+++ backports-6.1-9a0dddfb3/drivers/net/wireless/ath/ath12k/dp.c
@@ -1456,7 +1456,7 @@ int ath12k_dp_link_desc_init(struct ath12k_base *ab,
 	struct wbm_link_desc *desc;
 	u16 link_desc_size;
 	u32 tot_mem_sz;
-	u32 paddr;
+	u64 paddr;
 	int i, ret;
 	u32 cookie;
